log: Make it explicit an output got truncated
This will result in explicit truncation of values.

One thing I'm wondering: should truncation happen in the formatter?

With the current setup, when using the JSON formatter, a value with an object/array that is under MaxSize will output as an object. However, when it exceedts the MaxSize, the logged value will be a string.

This would be particularly tricky when people use an ELK stack, [since the inconsistent field types would result in dropped loglines](https://discuss.elastic.co/t/logs-drop-on-field-conflict/129259/3). We might want to treat truncation differently in JSON vs. logfmt.

On the other hand, that's part of the risk of logging JSON which doesn't conform a very clear schema.