log: Text formatter outputs logfmt (#24)
* log: Text formatter outputs logfmt

* Make text format use logfmt

Encode time, severity and message as k/v pairs.

* log: Add log.Fatal function (#20)

closes #18

* log: Rename Flush -> FlushAndDisableBuffering (#22)

The fact a function called Flush was also disabling buffering was
confusing.

* log: Make it explicit an output got truncated (#23)

This will result in explicit truncation of values.

One thing I'm wondering: should truncation happen in the formatter?

With the current setup, when using the JSON formatter, a value with an object/array that is under MaxSize will output as an object. However, when it exceedts the MaxSize, the logged value will be a string.

This would be particularly tricky when people use an ELK stack, [since the inconsistent field types would result in dropped loglines](https://discuss.elastic.co/t/logs-drop-on-field-conflict/129259/3). We might want to treat truncation differently in JSON vs. logfmt.

On the other hand, that's part of the risk of logging JSON which doesn't conform a very clear schema.

Co-authored-by: Raphael Simon <simon.raphael@gmail.com>

* Add `Fatal` description

* Update to latest

* Update FormatText comment to match latest

Co-authored-by: Raphael Simon <simon.raphael@gmail.com>