*BREAKING* Add ability to set pinger request path
This commit adds the ability to override the default path used by the
pinger to make health check requests to downstream dependencies.
It introduces options when creating a pinger and makes the scheme an
option. As a result the signature of the `NewPinger` function is now
different:
Before:
```
NewPinger("foo", "http", fooAddr)
```
Now:
```
NewPinger("foo", fooAddr)
```
or
```
NewPinger("foo", fooAddr, health.WithScheme("https"))
```