Add StartTrace and ContinueRemoteTrace
These functions make it possible to trace requests initiated
by code that do not run as part of a service, for example
workers.

StartTrace creates a new trace and initializes the context so
that the trace gets propagated to remote services when made
via tracing HTTP or gRPC clients.

ContinueRemoteTrace behaves similarly to StartTrace but
accepts an existing trace ID as argument so that traces can
be continued asynchronously (for example by an adapter
reading traced messages from a queueing system).