StartTrace and ContinueRemoteTrace (#95)
* Add StartTrace and ContinueRemoteTrace

These functions make it possible to trace requests initiated
by code that do not run as part of a service, for example
workers.

StartTrace creates a new trace and initializes the context so
that the trace gets propagated to remote services when made
via tracing HTTP or gRPC clients.

ContinueRemoteTrace behaves similarly to StartTrace but
accepts an existing trace ID as argument so that traces can
be continued asynchronously (for example by an adapter
reading traced messages from a queueing system).

* Consistently log trace IDs in debug mode

* Do not automatically log trace IDs

Let the client decide how and whether to log them. Typically they get
added to the log context when present (e.g. via the trace.Log function).

* Fix tests