Fix Trace Stream interceptors when `RecvWithContext` returns an error (#557)
* Fix Trace Stream interceptors when `RecvWithContext` returns an error

The implementation of the Trace Stream interceptors' `RecvWithContext`
handling did not take into account an error return where the message
(result or payload) would be `nil`. This change changes the way
`RecvWithContext` calls are intercepted so that if the `next` returns an
error, the interceptor will populate the receive context using the
passed in context and not even attempt to gather any OTel trace metadata
from a message (which usually would not be available in the error case,
but may be possible if another Goa interceptor were in the chain).

I don't believe it is worthwhile to support a case where an error is
returned by `next` along with a non-`nil` message, but if we did, we
would need to resort to using reflection due to the current state of Go
generics and `nil` checking:

```go
func traceStreamRecv[Message TraceStreamStreamingRecvMessage](
    ctx context.Context,
    info goa.InterceptorInfo,
    next goa.Endpoint,
    streamingMessage func(any) Message,
) (any, error) {
    msg, err := next(ctx, info.RawPayload())
    propagator := otel.GetTextMapPropagator()
    sm := streamingMessage(msg)
    rc, ok :=
ctx.Value(traceStreamRecvContextKey).(*traceStreamRecvContext)
    if !ok {
        panic(fmt.Errorf("clue interceptors trace stream receive method
called without prior setup (service: %v, method: %v)", info.Service(),
info.Method()))
    }
    if reflect.ValueOf(sm).Elem().IsZero() {
        rc.ctx = ctx
    } else {
        rc.ctx = propagator.Extract(ctx,
propagation.MapCarrier(sm.TraceMetadata()))
    }
    return msg, err
}
```

* Return `nil` instead of `err` since it is unnecessary